+++ Packages/webxr/Hidden~/WebGLTemplates/WebXR2020/webxr.js
@@ -902,7 +902,7 @@
             leftRect.w = (viewport.width / glLayer.framebufferWidth) * (glLayer.framebufferWidth / this.canvas.width);
             leftRect.h = (viewport.height / glLayer.framebufferHeight) * (glLayer.framebufferHeight / this.canvas.height);
           }
-        } else if (view.eye === 'right') {
+        } else if (view.eye === 'right' && viewport.width != 0 && viewport.height != 0) {
           eyeCount = 2;
           if (viewport) {
             rightRect.x = (viewport.x / glLayer.framebufferWidth) * (glLayer.framebufferWidth / this.canvas.width);