+++ Packages/webxr/CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Better handling for controllers buttons in the editor.
 - WebXRController TryUpdateButtons is now private.
 
+### Fixed
+- Disable controllers on XR session end.
+
 ## [0.4.1] - 2020-11-28
 ### Added
 - Support for Unity 2020.1.