+++ Packages/webxr/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - WebXRController Button Up/Down state.
+- Protect compilation on other platforms.
 
 ## [0.5.0] - 2020-12-19
 ### Added