+++ Packages/webxr/CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Multi touch support for Handheld AR.
+
+### Fixed
+- OnDisable in WebXRController.
+- Depth and Stencil clear issue in Handheld AR.
+- Ugly hack to fix WebXR Viewer viewports on iOS.
 
 ## [0.5.1] - 2020-12-26
 ### Fixed